% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{pls_assess}
\alias{pls_assess}
\title{Assess measurement and structural results}
\usage{
pls_assess(fit)
}
\arguments{
\item{fit}{Fitted model.}
}
\value{
Combined assessment list.
}
\description{
Assess measurement and structural results
}
