% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{pls_backend_object}
\alias{pls_backend_object}
\title{Extract the raw backend object}
\usage{
pls_backend_object(fit)
}
\arguments{
\item{fit}{A fitted model.}
}
\value{
Raw backend result or native result list.
}
\description{
Extract the raw backend object
}
