% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{pls_boot_ci}
\alias{pls_boot_ci}
\title{Extract bootstrap intervals}
\usage{
pls_boot_ci(boot, component = NULL)
}
\arguments{
\item{boot}{Bootstrap object.}

\item{component}{Optional component: \code{path}, \code{loading}, or \code{weight}.}
}
\value{
Data frame of intervals.
}
\description{
Extract bootstrap intervals
}
