% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_cluster}
\alias{pls_cluster}
\title{Define a cluster variable for multilevel extensions}
\usage{
pls_cluster(cluster)
}
\arguments{
\item{cluster}{Name of the cluster column in the data.}
}
\value{
A cluster relation specification.
}
\description{
Define a cluster variable for multilevel extensions
}
