% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{pls_compare_external}
\alias{pls_compare_external}
\title{Compare native parameters with external results}
\usage{
pls_compare_external(fit, external)
}
\arguments{
\item{fit}{Native fitted model.}

\item{external}{Data frame from \code{pls_import_external()}.}
}
\value{
Parameter comparison table.
}
\description{
Compare native parameters with external results
}
