% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{pls_csem_assess}
\alias{pls_csem_assess}
\title{Run cSEM post-estimation assessment}
\usage{
pls_csem_assess(fit, ...)
}
\arguments{
\item{fit}{A \code{plssem_fit} estimated with cSEM.}

\item{...}{Passed to \code{cSEM::assess()}.}
}
\value{
cSEM assessment object.
}
\description{
Run cSEM post-estimation assessment
}
