% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{pls_export_external}
\alias{pls_export_external}
\title{Export a model/data bundle for external comparison}
\usage{
pls_export_external(model, data, dir = "plssem_external_bundle")
}
\arguments{
\item{model}{Model specification.}

\item{data}{Data frame.}

\item{dir}{Destination directory.}
}
\value{
Paths to exported files.
}
\description{
Export a model/data bundle for external comparison
}
