% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pls_export_xlsx}
\alias{pls_export_xlsx}
\title{Export a table to XLSX}
\usage{
pls_export_xlsx(table, file)
}
\arguments{
\item{table}{Data frame.}

\item{file}{Output file.}
}
\description{
Export a table to XLSX
}
