% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_higher_order}
\alias{pls_higher_order}
\title{Define a higher-order construct}
\usage{
pls_higher_order(
  name,
  dimensions,
  type = c("reflective", "composite"),
  approach = c("two_stage", "mixed", "repeated")
)
}
\arguments{
\item{name}{Higher-order construct name.}

\item{dimensions}{Names of lower-order constructs.}

\item{type}{Either \code{reflective} or \code{composite}.}

\item{approach}{Higher-order estimation strategy.}
}
\value{
A higher-order construct specification.
}
\description{
Define a higher-order construct
}
