% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_measurement}
\alias{pls_measurement}
\title{Combine measurement specifications}
\usage{
pls_measurement(...)
}
\arguments{
\item{...}{Objects created by \code{pls_reflective()}, \code{pls_composite()},
\code{pls_formative()}, or \code{pls_higher_order()}.}
}
\value{
A measurement model.
}
\description{
Combine measurement specifications
}
