% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{pls_measurement_assess}
\alias{pls_measurement_assess}
\title{Assess the measurement model}
\usage{
pls_measurement_assess(fit)
}
\arguments{
\item{fit}{Fitted model.}
}
\value{
List with loadings, reliability, AVE, HTMT, and formative VIF where applicable.
}
\description{
Assess the measurement model
}
