% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_mediation}
\alias{pls_mediation}
\title{Define mediation}
\usage{
pls_mediation(from, via, to, direct = TRUE)
}
\arguments{
\item{from}{Antecedent construct.}

\item{via}{Mediator construct.}

\item{to}{Outcome construct.}

\item{direct}{Include the direct path from antecedent to outcome.}
}
\value{
A mediation specification.
}
\description{
Define mediation
}
