% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{pls_micom}
\alias{pls_micom}
\title{Run MICOM using cSEM}
\usage{
pls_micom(fit, R = 4999L, seed = NULL, ...)
}
\arguments{
\item{fit}{A cSEM multigroup fit.}

\item{R}{Number of permutation runs selected by the user.}

\item{seed}{Optional seed.}

\item{...}{Additional backend arguments.}
}
\value{
MICOM result.
}
\description{
Run MICOM using cSEM
}
