% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_model}
\alias{pls_model}
\title{Build a complete PLS-SEM specification}
\usage{
pls_model(measurement, structural, name = NULL)
}
\arguments{
\item{measurement}{Measurement model.}

\item{structural}{Structural model.}

\item{name}{Optional model name.}
}
\value{
A \code{plssem_model} object.
}
\description{
Build a complete PLS-SEM specification
}
