% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_moderation}
\alias{pls_moderation}
\title{Define moderation}
\usage{
pls_moderation(
  predictor,
  moderator,
  outcome,
  method = c("two_stage", "product_indicator", "orthogonal")
)
}
\arguments{
\item{predictor}{Predictor construct.}

\item{moderator}{Moderator construct.}

\item{outcome}{Outcome construct.}

\item{method}{Interaction construction method.}
}
\value{
A moderation specification.
}
\description{
Define moderation
}
