% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_path}
\alias{pls_path}
\title{Define a direct structural path}
\usage{
pls_path(from, to)
}
\arguments{
\item{from}{Predictor construct.}

\item{to}{Outcome construct.}
}
\value{
A structural relation specification.
}
\description{
Define a direct structural path
}
