% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_quadratic}
\alias{pls_quadratic}
\title{Define a quadratic structural effect}
\usage{
pls_quadratic(predictor, outcome)
}
\arguments{
\item{predictor}{Predictor construct.}

\item{outcome}{Outcome construct.}
}
\value{
A quadratic-effect specification.
}
\description{
Define a quadratic structural effect
}
