% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_reflective}
\alias{pls_reflective}
\title{Define a reflective construct}
\usage{
pls_reflective(name, indicators)
}
\arguments{
\item{name}{Construct name.}

\item{indicators}{Character vector of indicator names.}
}
\value{
A measurement-construct specification.
}
\description{
Define a reflective construct
}
