% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pls_save_plot}
\alias{pls_save_plot}
\title{Save a publication figure}
\usage{
pls_save_plot(plot, file, width = 8, height = 6, dpi = 600)
}
\arguments{
\item{plot}{ggplot object.}

\item{file}{Output filename.}

\item{width, height}{Size in inches.}

\item{dpi}{Raster resolution.}
}
\description{
Save a publication figure
}
