% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{pls_scores}
\alias{pls_scores}
\title{Extract latent/composite scores}
\usage{
pls_scores(fit)
}
\arguments{
\item{fit}{A fitted model.}
}
\value{
Data frame of construct scores when available.
}
\description{
Extract latent/composite scores
}
