% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_structural}
\alias{pls_structural}
\title{Combine structural relations}
\usage{
pls_structural(...)
}
\arguments{
\item{...}{Structural relation specifications.}
}
\value{
A structural model.
}
\description{
Combine structural relations
}
