% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{pls_structural_assess}
\alias{pls_structural_assess}
\title{Assess the structural model}
\usage{
pls_structural_assess(fit)
}
\arguments{
\item{fit}{Fitted model.}
}
\value{
List with paths, R2, f2 effect sizes, and construct-score VIF.
}
\description{
Assess the structural model
}
