% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{pls_syntax}
\alias{pls_syntax}
\title{Convert a specification to backend syntax}
\usage{
pls_syntax(model, higher_order = TRUE, dialect = c("cSEM", "plssem", "lavaan"))
}
\arguments{
\item{model}{A \code{plssem_model}.}

\item{higher_order}{Include higher-order definitions.}

\item{dialect}{Syntax dialect: \code{cSEM}, \code{plssem}, or \code{lavaan}.}
}
\value{
Character string containing model syntax.
}
\description{
Convert a specification to backend syntax
}
