% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour_report.R
\name{pls_tour}
\alias{pls_tour}
\title{Guided PLS-SEM teaching tour}
\usage{
pls_tour(
  topic = c("foundations", "measurement", "structural", "bootstrap", "mediation",
    "moderation", "prediction", "multigroup", "nonlinear", "multilevel", "bayesian",
    "reviewer")
)
}
\arguments{
\item{topic}{Topic name.}
}
\value{
Ordered teaching steps and suggested functions.
}
\description{
Guided PLS-SEM teaching tour
}
