% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{plsSEMflow-package}
\alias{plsSEMflow}
\alias{plsSEMflow-package}
\title{plsSEMflow: Integrated Partial Least Squares Structural Equation Modeling}
\description{
An R-first integrator for PLS-SEM, bootstrap inference, model assessment,
prediction, advanced optional backends, publication-ready outputs, and
agronomic teaching workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wep69/plsSEMflow}
  \item Report bugs at \url{https://github.com/wep69/plsSEMflow/issues}
}

}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
