% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.plssem_bootstrap}
\alias{print.plssem_bootstrap}
\title{Print bootstrap results}
\usage{
\method{print}{plssem_bootstrap}(x, ...)
}
\arguments{
\item{x}{Bootstrap object.}

\item{...}{Unused.}
}
\description{
Print bootstrap results
}
