% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{print.plssem_external_workflow}
\alias{print.plssem_external_workflow}
\title{Print an external specialist workflow descriptor}
\usage{
\method{print}{plssem_external_workflow}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_external_workflow} object.}

\item{...}{Unused.}
}
\description{
Print an external specialist workflow descriptor
}
