% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.plssem_fit}
\alias{print.plssem_fit}
\title{Print a fitted PLS-SEM model}
\usage{
\method{print}{plssem_fit}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_fit}.}

\item{...}{Unused.}
}
\description{
Print a fitted PLS-SEM model
}
