% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{print.plssem_heterogeneity}
\alias{print.plssem_heterogeneity}
\title{Print a heterogeneity result}
\usage{
\method{print}{plssem_heterogeneity}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_heterogeneity} object.}

\item{...}{Unused.}
}
\description{
Print a heterogeneity result
}
