% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_output.R
\name{print.plssem_mga}
\alias{print.plssem_mga}
\title{Print native MGA result}
\usage{
\method{print}{plssem_mga}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_mga} object.}

\item{...}{Unused.}
}
\description{
Print native MGA result
}
