% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{print.plssem_model}
\alias{print.plssem_model}
\title{Print a PLS-SEM model specification}
\usage{
\method{print}{plssem_model}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_model}.}

\item{...}{Unused.}
}
\description{
Print a PLS-SEM model specification
}
