% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{print.plssem_prediction}
\alias{print.plssem_prediction}
\title{Print prediction results}
\usage{
\method{print}{plssem_prediction}(x, ...)
}
\arguments{
\item{x}{Prediction object.}

\item{...}{Unused.}
}
\description{
Print prediction results
}
