% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{print.plssem_validation_case}
\alias{print.plssem_validation_case}
\title{Print a simulation validation case}
\usage{
\method{print}{plssem_validation_case}(x, ...)
}
\arguments{
\item{x}{A \code{plssem_validation_case}.}

\item{...}{Unused.}
}
\description{
Print a simulation validation case
}
