% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{pls_cvpat}
\alias{pls_cvpat}
\title{Run cSEM CVPAT}
\usage{
pls_cvpat(fit, ...)
}
\arguments{
\item{fit}{A cSEM-backed fit.}

\item{...}{Passed to \code{cSEM::testCVPAT()}.}
}
\value{
Backend test result.
}
\description{
Run cSEM CVPAT
}
