% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{pls_mgd}
\alias{pls_mgd}
\title{Test multigroup differences using cSEM}
\usage{
pls_mgd(fit, ...)
}
\arguments{
\item{fit}{A cSEM multigroup fit.}

\item{...}{Passed to \code{cSEM::testMGD()}.}
}
\value{
Multigroup-difference result.
}
\description{
Test multigroup differences using cSEM
}
